implement screen grid

+++ d12util/Shader.h
@@ -19,7 +19,7 @@ class Shader : NonCopyable
 
 public:
     const D3D12_INPUT_ELEMENT_DESC *inputLayout() const { return m_layout.data(); }
-    int inputLayoutCount() const { return m_layout.size(); }
+    int inputLayoutCount() const { return (int)m_layout.size(); }
 
     bool Initialize(const ComPtr<ID3D12Device> &device,
                     const ComPtr<ID3D12RootSignature> &rootSignature,
